Refactor useThemeImage hook location and update imports across components
