+++ src/components/layouts/ecosystem/iqsocial.tsx
@@ -1,6 +1,6 @@
 "use client";
 import InViewAnimateBottom from "@/components/transitions/InViewAnimateBottom";
-import useThemeImage from "@/modules/helpers/useThemeImage";
+import useThemeImage from "@/hooks/useThemeImage";
 import Image from "next/image";
 import React from "react";
 import { TokenBrief } from "../Iqtoken";
